Move the `Project` class to `Dock\Compose` namespace
